refactor: add schemas namespace
Schemas aren't just used in operations, they can be used elsewhere
so I've moved them into their own schemas namespace.
