+++ src/Operations/MediaType.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace OpenApiSchema\Operations;
 
+use OpenApiSchema\Schema\Schema;
 use OpenApiSchema\Utils\Marshallable;
 use OpenApiSchema\Utils\HasCustomAttributes;
 use OpenApiSchema\Utils\ConvertsSelfToMarshallable;
